Adds scaladocs and removes some options that made no sense in the current implementation
