+++ akka-zeromq/src/main/scala/akka/zeromq/ZMQMessageDeserializer.scala
@@ -3,9 +3,16 @@
  */
 package akka.zeromq
 
+/**
+ * A single message frame of a zeromq message
+ * @param payload
+ */
 case class Frame(payload: Seq[Byte])
 object Frame { def apply(s: String): Frame = Frame(s.getBytes) }
 
+/**
+ * Deserializes ZeroMQ messages into an immutable sequence of frames
+ */
 class ZMQMessageDeserializer extends Deserializer {
   def apply(frames: Seq[Frame]) = ZMQMessage(frames)
 }
