+++ akka-zeromq/src/main/scala/akka/zeromq/ZeroMQExtension.scala
@@ -6,12 +6,21 @@ package akka.zeromq
 import akka.util.duration._
 import org.zeromq.{ ZMQ ⇒ JZMQ }
 import akka.actor._
-import akka.dispatch.{ Dispatcher, Await }
+import akka.dispatch.{ Await }
 
+/**
+ * A Model to represent a version of the zeromq library
+ * @param major
+ * @param minor
+ * @param patch
+ */
 case class ZeroMQVersion(major: Int, minor: Int, patch: Int) {
   override def toString = "%d.%d.%d".format(major, minor, patch)
 }
 
+/**
+ * The [[akka.actor.ExtensionId]] and [[akka.actor.ExtensionIdProvider]] for the ZeroMQ module
+ */
 object ZeroMQExtension extends ExtensionId[ZeroMQExtension] with ExtensionIdProvider {
   def lookup() = this
   def createExtension(system: ActorSystemImpl) = new ZeroMQExtension(system)
@@ -25,25 +34,49 @@ object ZeroMQExtension extends ExtensionId[ZeroMQExtension] with ExtensionIdProv
     }
   }
 }
+
+/**
+ * The extension for the ZeroMQ module
+ *
+ * @param system The ActorSystem this extension belongs to.
+ */
 class ZeroMQExtension(system: ActorSystem) extends Extension {
 
+  /**
+   * The version of the ZeroMQ library
+   * @return a [[akka.zeromq.ZeroMQVersion]]
+   */
   def version = {
     ZeroMQVersion(JZMQ.getMajorVersion, JZMQ.getMinorVersion, JZMQ.getPatchVersion)
   }
 
+  /**
+   * Factory method to create the [[akka.actor.Props]] to build the ZeroMQ socket actor.
+   *
+   * @param socketParameters a varargs list of [[akka.zeromq.SocketOption]] to configure the socket
+   * @return the [[akka.actor.Props]]
+   */
   def newSocketProps(socketParameters: SocketOption*): Props = {
     verifyZeroMQVersion
     require(ZeroMQExtension.check[SocketType.ZMQSocketType](socketParameters), "A socket type is required")
     Props(new ConcurrentSocketActor(socketParameters)).withDispatcher("akka.zeromq.socket-dispatcher")
   }
 
+  /**
+   * Factory method to create the actor representing the ZeroMQ socket.
+   * You can pass in as many configuration options as you want and the order of the configuration options doesn't matter
+   * They are matched on type and the first one found wins.
+   *
+   * @param socketParameters a varargs list of [[akka.zeromq.SocketOption]] to configure the socke
+   * @return
+   */
   def newSocket(socketParameters: SocketOption*): ActorRef = {
     implicit val timeout = system.settings.ActorTimeout
-    val req = (zeromq ? newSocketProps(socketParameters: _*)).mapTo[ActorRef]
+    val req = (zeromqGuardian ? newSocketProps(socketParameters: _*)).mapTo[ActorRef]
     Await.result(req, timeout.duration)
   }
 
-  val zeromq: ActorRef = {
+  private val zeromqGuardian: ActorRef = {
     verifyZeroMQVersion
     system.actorOf(Props(new Actor {
       def receive = { case p: Props ⇒ sender ! context.actorOf(p) }
