+++ akka-zeromq/src/main/scala/package.scala
@@ -2,10 +2,25 @@ package akka
 
 import actor.ActorSystem
 
+/**
+ * A package object with an implicit conversion for the actor system as a convenience
+ */
 package object zeromq {
-  implicit def zeromqSystem(system: ActorSystem) = new {
-    def zeromq = system.extension(ZeroMQExtension)
-  }
 
+  /**
+   * Creates a zeromq actor system implicitly
+   * @param system
+   * @return An augmented [[akka.actor.ActorSystem]]
+   */
+  implicit def zeromqSystem(system: ActorSystem) = system.extension(ZeroMQExtension)
+
+  /**
+   * Convenience accessor to subscribe to all events
+   */
   val SubscribeAll = Subscribe(Seq.empty)
+
+  /**
+   * Set the linger to 0, doesn't block and discards messages that haven't been sent yet.
+   */
+  val NoLinger = Linger(0)
 }
